+++ jobs-spring-boot-starter/src/main/java/com/baomidou/jobs/starter/router/strategy/ExecutorRouteConsistentHash.java
@@ -57,7 +57,7 @@ public class ExecutorRouteConsistentHash implements IExecutorRouter {
         return truncateHashCode;
     }
 
-    public String hashJob(int jobId, List<String> addressList) {
+    public String hashJob(Long jobId, List<String> addressList) {
 
         // ------A1------A2-------A3------
         // -----------J1------------------
